+++ src/videojs-contrib-hls.js
@@ -16,8 +16,9 @@ import videojs from 'video.js';
 import { MasterPlaylistController } from './master-playlist-controller';
 import Config from './config';
 import renditionSelectionMixin from './rendition-mixin';
-import GapSkipper from './gap-skipper';
 import window from 'global/window';
+import PlaybackWatcher from './playback-watcher';
+import reloadSourceOnError from './reload-source-on-error';
 
 const Hls = {
   PlaylistLoader,
@@ -371,7 +372,10 @@ class HlsHandler extends Component {
     this.options_.tech = this.tech_;
     this.options_.externHls = Hls;
     this.masterPlaylistController_ = new MasterPlaylistController(this.options_);
-    this.gapSkipper_ = new GapSkipper(this.options_);
+    this.playbackWatcher_ = new PlaybackWatcher(
+      videojs.mergeOptions(this.options_, {
+        seekable: () => this.seekable()
+      }));
 
     // `this` in selectPlaylist should be the HlsHandler for backwards
     // compatibility with < v2
@@ -502,10 +506,12 @@ class HlsHandler extends Component {
   * Abort all outstanding work and cleanup.
   */
   dispose() {
+    if (this.playbackWatcher_) {
+      this.playbackWatcher_.dispose();
+    }
     if (this.masterPlaylistController_) {
       this.masterPlaylistController_.dispose();
     }
-    this.gapSkipper_.dispose();
     this.tech_.audioTracks().removeEventListener('change', this.audioTrackChange_);
     super.dispose();
   }
@@ -657,6 +663,7 @@ videojs.Hls = Hls;
 videojs.m3u8 = m3u8;
 videojs.registerComponent('Hls', Hls);
 videojs.options.hls = videojs.options.hls || {};
+videojs.plugin('reloadSourceOnError', reloadSourceOnError);
 
 module.exports = {
   Hls,